authorization: Move HTTP API types in submodule
